+++ Components/Main/Playlist.py
@@ -5,7 +5,7 @@ from Utils.utils import spotifyGetAPI
 class Playlist(Menu):
 
     def __init__(self, name, playlistId):
-        response = utils.spotifyGetAPI(f"/playlists/{playlistId}/tracks", cache=True, paged=True)
+        response = spotifyGetAPI(f"/playlists/{playlistId}/tracks", cache=True, paged=True)
         items = []
         for track in response:
             items.append(SongItem(track, f"spotify:playlist:{playlistId}"))
