+++ Components/Main/PlaylistMenu.py
@@ -19,7 +19,7 @@ class PlaylistMenu(Menu):
     def __init__(self):
         name = "Playlist Menu"
         items = []
-        response = utils.spotifyGetAPI("/me/playlists", cache=True, paged=True)
+        response = spotifyGetAPI("/me/playlists", cache=True, paged=True)
         for playlist in response:
             items.append(PlaylistItem(playlist['name'], playlist['id']))
         super().__init__(name, items)
